requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
